refactor(step4): split class by responsibility
